backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
openai==1.3.7
python-decouple==3.8
Pillow==10.1.0 